Reference filtering now works
