+++ argonaut/src/main/resources/application.properties
@@ -32,11 +32,11 @@ conformance.security.token-endpoint=unset
 conformance.security.authorize-endpoint=unset
 conformance.security.description=http://docs.smarthealthit.org/
 conformance.resourceDocumentation=Implemented per specification. See http://hl7.org/fhir/DSTU2/http.html
-include.appointment=true
-include.encounter=true
-include.location=false
-include.organization=true
-include.practitioner=true
+included-references.appointment=true
+included-references.encounter=true
+included-references.location=true
+included-references.organization=true
+included-references.practitioner=true
